+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add strict secret/key requirements check to the "JSON Web Token (JWT) Decoder/Encoder" tool.
+- Add gutter icon with readable a readable date/time format for UNIX timestamps in the "JSON Web Token (JWT) Decoder/Encoder" tool.
 - The "JSON Web Token (JWT) Decoder/Encoder" tool interface now includes sliders to flexibly change the size of the editors.
 
 ### Changed